Move ActionResponse to correct home, add docs
