+++ ScriptedEvents/Actions/Interfaces/IAction.cs
@@ -1,5 +1,8 @@
 ﻿namespace ScriptedEvents.Actions.Interfaces
 {
+    /// <summary>
+    /// Represents any action.
+    /// </summary>
     public interface IAction
     {
         string Name { get; }
