+++ ScriptedEvents/Actions/Interfaces/IHiddenAction.cs
@@ -1,5 +1,8 @@
 ﻿namespace ScriptedEvents.Actions.Interfaces
 {
+    /// <summary>
+    /// Represents an action that does not show in the "HELP LIST" action.
+    /// </summary>
     public interface IHiddenAction
     {
     }
