+++ ScriptedEvents/ScriptedEvents.csproj
@@ -78,7 +78,7 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Actions\Interfaces\ActionResponse.cs" />
+    <Compile Include="Structures\ActionResponse.cs" />
     <Compile Include="Actions\Interfaces\ICustomReadDisplay.cs" />
     <Compile Include="Actions\Interfaces\IHelpInfo.cs" />
     <Compile Include="Actions\Interfaces\IHiddenAction.cs" />
